apps/ocr-worker-python/requirements.txt
redis==5.0.1
Pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
python-docx==1.1.0
python-multipart==0.0.6
